Move platform/mediastream/mac to ../cocoa https://bugs.webkit.org/show_bug.cgi?id=310255
Reviewed by Youenn Fablet.

Make it more self-evident which files are actually shared across ports.

Canonical link: https://commits.webkit.org/309543@main
