+++ Source/WebCore/SourcesCocoa.txt
@@ -661,35 +661,35 @@ platform/mediarecorder/MediaRecorderPrivateAVFImpl.cpp
 platform/mediarecorder/MediaRecorderPrivateEncoder.cpp
 platform/mediarecorder/cocoa/MediaRecorderPrivateWriterAVFObjC.mm @nonARC
 platform/mediarecorder/cocoa/MediaRecorderPrivateWriterWebM.cpp
+platform/mediastream/cocoa/AVCaptureDeviceManager.mm @nonARC
+platform/mediastream/cocoa/AVVideoCaptureSource.mm @nonARC @no-unify
 platform/mediastream/cocoa/AudioMediaStreamTrackRendererCocoa.cpp
 platform/mediastream/cocoa/AudioMediaStreamTrackRendererInternalUnit.cpp
 platform/mediastream/cocoa/AudioMediaStreamTrackRendererUnit.cpp
+platform/mediastream/cocoa/BaseAudioCaptureUnit.cpp
+platform/mediastream/cocoa/CoreAudioCaptureDevice.cpp
+platform/mediastream/cocoa/CoreAudioCaptureDeviceManager.cpp
+platform/mediastream/cocoa/CoreAudioCaptureSource.cpp
+platform/mediastream/cocoa/CoreAudioCaptureUnit.cpp
+platform/mediastream/cocoa/DisplayCaptureManagerCocoa.cpp
 platform/mediastream/cocoa/DisplayCaptureSourceCocoa.cpp
 platform/mediastream/cocoa/IncomingAudioMediaStreamTrackRendererUnit.cpp
+platform/mediastream/cocoa/MediaStreamTrackAudioSourceProviderCocoa.cpp
+platform/mediastream/cocoa/MockAudioCaptureUnit.mm @nonARC
+platform/mediastream/cocoa/MockRealtimeVideoSourceCocoa.mm @nonARC
+platform/mediastream/cocoa/RealtimeIncomingAudioSourceCocoa.cpp
+platform/mediastream/cocoa/RealtimeIncomingVideoSourceCocoa.mm @nonARC
+platform/mediastream/cocoa/RealtimeMediaSourceCenterCocoa.cpp
+platform/mediastream/cocoa/RealtimeMediaSourceCenterCocoa.mm @nonARC
+platform/mediastream/cocoa/RealtimeOutgoingAudioSourceCocoa.cpp
+platform/mediastream/cocoa/RealtimeOutgoingVideoSourceCocoa.cpp
+platform/mediastream/cocoa/ScreenCaptureKitCaptureSource.mm @nonARC
+platform/mediastream/cocoa/ScreenCaptureKitSharingSessionManager.mm @nonARC
+platform/mediastream/cocoa/WebAudioSourceProviderCocoa.mm @nonARC
 platform/mediastream/ios/AVAudioSessionCaptureDevice.mm @nonARC
 platform/mediastream/ios/AVAudioSessionCaptureDeviceManager.mm @nonARC @no-unify
 platform/mediastream/ios/CoreAudioCaptureSourceIOS.mm @nonARC
 platform/mediastream/libwebrtc/LibWebRTCProviderCocoa.cpp
-platform/mediastream/mac/AVCaptureDeviceManager.mm @nonARC
-platform/mediastream/mac/AVVideoCaptureSource.mm @nonARC @no-unify
-platform/mediastream/mac/BaseAudioCaptureUnit.cpp
-platform/mediastream/mac/CoreAudioCaptureDevice.cpp
-platform/mediastream/mac/CoreAudioCaptureDeviceManager.cpp
-platform/mediastream/mac/CoreAudioCaptureSource.cpp
-platform/mediastream/mac/CoreAudioCaptureUnit.cpp
-platform/mediastream/mac/DisplayCaptureManagerCocoa.cpp
-platform/mediastream/mac/MediaStreamTrackAudioSourceProviderCocoa.cpp
-platform/mediastream/mac/MockAudioCaptureUnit.mm @nonARC
-platform/mediastream/mac/MockRealtimeVideoSourceMac.mm @nonARC
-platform/mediastream/mac/RealtimeIncomingAudioSourceCocoa.cpp
-platform/mediastream/mac/RealtimeIncomingVideoSourceCocoa.mm @nonARC
-platform/mediastream/mac/RealtimeMediaSourceCenterMac.cpp
-platform/mediastream/mac/RealtimeMediaSourceCenterMac.mm @nonARC
-platform/mediastream/mac/RealtimeOutgoingAudioSourceCocoa.cpp
-platform/mediastream/mac/RealtimeOutgoingVideoSourceCocoa.cpp
-platform/mediastream/mac/ScreenCaptureKitCaptureSource.mm @nonARC
-platform/mediastream/mac/ScreenCaptureKitSharingSessionManager.mm @nonARC
-platform/mediastream/mac/WebAudioSourceProviderCocoa.mm @nonARC
 platform/mock/MediaPlaybackTargetMock.cpp
 platform/mock/MediaPlaybackTargetPickerMock.cpp
 platform/mock/MockAudioDestinationCocoa.cpp
