+++ Source/WebCore/platform/mediastream/cocoa/AVCaptureDeviceManager.h
