+++ Source/WebCore/platform/mediastream/cocoa/AVVideoCaptureSource.h
