+++ Source/WebCore/platform/mediastream/cocoa/BaseAudioCaptureUnit.h
