+++ Source/WebCore/platform/mediastream/cocoa/CoreAudioCaptureDevice.cpp
