+++ Source/WebCore/platform/mediastream/cocoa/CoreAudioCaptureDevice.h
