+++ Source/WebCore/platform/mediastream/cocoa/CoreAudioCaptureDeviceManager.cpp
