+++ Source/WebCore/platform/mediastream/cocoa/CoreAudioCaptureDeviceManager.h
