+++ Source/WebCore/platform/mediastream/cocoa/CoreAudioCaptureSource.cpp
@@ -387,10 +387,10 @@ void CoreAudioCaptureSource::handleNewCurrentMicrophoneDevice(const CaptureDevic
         return;
 
     RELEASE_LOG_INFO(WebRTC, "CoreAudioCaptureSource switching from '%s' to '%s'", name().utf8().data(), device.label().utf8().data());
-    
+
     setName(AtomString { device.label() });
     setPersistentId(device.persistentId());
-    
+
     m_currentSettings = { };
     m_capabilities = { };
