+++ Source/WebCore/platform/mediastream/cocoa/CoreAudioCaptureSource.h
