+++ Source/WebCore/platform/mediastream/cocoa/CoreAudioCaptureUnit.h
