+++ Source/WebCore/platform/mediastream/cocoa/DisplayCaptureManagerCocoa.cpp
@@ -25,6 +25,7 @@
 
 #include "config.h"
 #include "DisplayCaptureManagerCocoa.h"
+
 #include "RealtimeMediaSourceCenter.h"
 
 #if ENABLE(MEDIA_STREAM)
