+++ Source/WebCore/platform/mediastream/cocoa/DisplayCaptureManagerCocoa.h
