+++ Source/WebCore/platform/mediastream/cocoa/MockAudioCaptureUnit.mm
