+++ Source/WebCore/platform/mediastream/cocoa/MockRealtimeVideoSourceCocoa.h
