+++ Source/WebCore/platform/mediastream/cocoa/RealtimeIncomingVideoSourceCocoa.mm
