+++ Source/WebCore/platform/mediastream/cocoa/RealtimeOutgoingAudioSourceCocoa.cpp
