+++ Source/WebCore/platform/mediastream/cocoa/RealtimeOutgoingAudioSourceCocoa.h
