+++ Source/WebCore/platform/mediastream/cocoa/RealtimeOutgoingVideoSourceCocoa.h
