+++ Source/WebCore/platform/mediastream/cocoa/RealtimeVideoUtilities.h
