+++ Source/WebCore/platform/mediastream/cocoa/ScreenCaptureKitCaptureSource.h
