+++ Source/WebCore/platform/mediastream/cocoa/ScreenCaptureKitCaptureSource.mm
