+++ Source/WebCore/platform/mediastream/cocoa/ScreenCaptureKitSharingSessionManager.h
