+++ Source/WebCore/platform/mediastream/cocoa/ScreenCaptureKitSharingSessionManager.mm
