+++ Source/WebCore/platform/mediastream/cocoa/WebAudioSourceProviderCocoa.h
