+++ Source/WebCore/platform/mediastream/cocoa/WebAudioSourceProviderCocoa.mm
